Check whether calendar_popup exists rather than awesome.version
It may happen that the version string on certain builds of awesome is **not**
"v.4.2" but just "4.2". This makes the existing check ineffective.

Moreover, the calendar API is likely to be kept in further releases of
Awesome. The current code would fail e.g. for version 4.2.1 or 4.3

It's much more effective checking whether the calendar API exists on
awful.widget rather than trying to check the version of awesome.